third_party/blink/web_tests/platform/generic/fast/js/typed-array-allocation-failure-expected.txt

This test checks that TypedArray allocation failures should throw

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new ArrayBuffer(7 * 1125899906842624) threw exception RangeError.
PASS new Uint8Array(7 * 1125899906842624) threw exception RangeError.
PASS new Uint8ClampedArray(7 * 1125899906842624) threw exception RangeError.
PASS new Uint16Array(7 * 1125899906842624) threw exception RangeError.
PASS new Uint32Array(7 * 1125899906842624) threw exception RangeError.
PASS new Int8Array(7 * 1125899906842624) threw exception RangeError.
PASS new Int16Array(7 * 1125899906842624) threw exception RangeError.
PASS new Int32Array(7 * 1125899906842624) threw exception RangeError.
PASS new Float32Array(7 * 1125899906842624) threw exception RangeError.
PASS new Float64Array(7 * 1125899906842624) threw exception RangeError.
PASS successfullyParsed is true

TEST COMPLETE
